requirements.txt
Django==4.2.7
gunicorn==20.0.4
djangorestframework==3.14.0
Pillow==10.1.0
mysql-connector-python==8.0.32
itsdangerous==2.1.2